changed main version to be threaded; added support for things with only fail2ban and no ufw
